chore: update priority type to float, use thread vector manage threads.

+++ Plugin~/WebRTCPlugin/Codec/IEncoder.h
@@ -24,7 +24,7 @@ namespace webrtc
         virtual void UpdateSettings() = 0;
         virtual bool CopyBuffer(void* frame) = 0;
         virtual bool EncodeFrame(int64_t timestamp_us) = 0;
-        virtual bool EncodeFrame(int64_t timestamp_us, uint32_t* priorityArray) = 0;
+        virtual bool EncodeFrame(int64_t timestamp_us, float* priorityArray) = 0;
         virtual bool IsSupported() const = 0;
         virtual void SetIdrFrame() = 0;
         virtual uint64 GetCurrentFrameCount() const = 0;
