+++ Plugin~/WebRTCPlugin/Codec/NvCodec/NvEncoder.h
@@ -52,7 +52,7 @@ namespace webrtc
         void UpdateSettings() override;
         bool CopyBuffer(void* frame) override;
         bool EncodeFrame(int64_t timestamp_us) override;
-        bool EncodeFrame(int64_t timestamp_us, uint32_t* priorityArray) override;
+        bool EncodeFrame(int64_t timestamp_us, float* priorityArray) override;
         bool IsSupported() const override { return m_isNvEncoderSupported; }
         void SetIdrFrame()  override { isIdrFrame = true; }
         uint64 GetCurrentFrameCount() const override { return frameCount; }
