+++ Plugin~/WebRTCPlugin/Codec/SoftwareCodec/SoftwareEncoder.h
@@ -21,7 +21,7 @@ namespace webrtc
         void UpdateSettings() override {}
         bool CopyBuffer(void* frame) override;
         bool EncodeFrame(int64_t timestamp_us) override;
-        bool EncodeFrame(int64_t timestamp_us, uint32_t* priorityArray) override;
+        bool EncodeFrame(int64_t timestamp_us, float* priorityArray) override;
         bool IsSupported() const override { return true; }
         void SetIdrFrame() override {}
         uint64 GetCurrentFrameCount() const override { return m_frameCount; }
